# back_end/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Recherche d'emploi",
        "description": "Documentation de l'API Laravel pour la plateforme de recherche d'emploi",
        "contact": {
            "email": "hariznabil663@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur principal"
        }
    ],
    "paths": {
        "/admin/users": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get all users with roles",
                "operationId": "06d892de062bd84601131f920878264e",
                "responses": {
                    "200": {
                        "description": "List of users with roles"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/personal-data": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get authenticated admin's personal data",
                "operationId": "b2034f54df4272a41ba94109a3ca0396",
                "responses": {
                    "200": {
                        "description": "Authenticated user's data"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/users/{id}": {
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Update a user's name and/or email",
                "operationId": "54b1af38b1aeff6584e7f13d9fcf34f6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "New Name"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "new@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Delete a user",
                "operationId": "c3d89b5334aaae1703053279305a83be",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted successfully"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/employer/applications": {
            "get": {
                "tags": [
                    "Applications"
                ],
                "summary": "Get applications for jobs posted by the authenticated employer",
                "operationId": "672ab5fe3bfb6da5a8fdce1959c6e26e",
                "responses": {
                    "200": {
                        "description": "List of applications for employer's jobs"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/applications": {
            "get": {
                "tags": [
                    "Applications"
                ],
                "summary": "Get all applications (Admin only)",
                "operationId": "4dac0d67b7b509131bafbd8db2c97870",
                "responses": {
                    "200": {
                        "description": "List of all job applications"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}